Bump to ver 0.9.0
I've added in preliminary code to convert MER files older than ver 5 into MER files that can be restored as a project file.

I've also added in versions of each functions that take a list of file paths and use rayon to run them in parallel. This is manly to try and give a performance boost for the Python bindings on a large number of files.