Added int he file_type function
Added in a rough file_type function to check if the file is an
MER or an APA, to better handle the differences between them.

Right now the check is crude, just looks at the file extension,
which since ApplicationManager ensures that, and it usually
matters in Windows, should be good enough, but I might try and
add in some better form of checking later.